requirements.txt
pytest-playwright==0.4.4
